+++ changelog.md
@@ -6,10 +6,13 @@
 * How to handle monk's **Metal Strikes** and **Adamantine Strikes**?
 * Oracle: Where is the **Frost**, **Mountain**, **Radiance** mysteries?
 * Should Thaumaturge class be included?
-## Update 0.0.2
+* Are we using the apprentice options?
+## Updates
+### v0.0.2
 * Added override to PlayerSheetPF2e to have ancestry, heritage, background, class and deity buttons open Dark Sun compendiums instead of the Core PF2E compendiums.
 * Fixed formatting of Sorcerer-Monarch description pages.
 * Fixed a few grammatical errors in different entry descriptions.
 * Fixed **Resourceful Thri-Kreen** accidently giving martial weapon proficiency.
 * Added **Level Zero** class, since PF2E doesn't handle Level 0 characters according to the core rulebooks.
-* Added script to automatically popup the module's introduction journal when a player enters the world for the first time.
+* Automatically popup the module's introduction journal when a player enters the world for the first time.
+* Automatically set Compendium Browser to only use Dark Sun sources.